Crate oak_functions_loader

Modules

grpc

gRPC server for Oak Functions.

logger

lookup_data

server

Structs

LoadLookupDataConfig

Configuration to load the LookupData.

Opt

Command line options specificing how to run the Oak Functions Runtime, which are set by the team operating Oak Functions Runtime as the platform.

Enums

Data

Functions

create_base_extension_factories

lib_main

This crate is just a library so this function does not get executed directly by anything, it needs to be wrapped in the “actual” main from a bin crate.

load_lookup_data

Creates LookupDataManager and sets up LookupDataRefresher.

Type Definitions

OakFunctionsBoxedExtensionFactory