Struct oak_functions_loader::logger::Logger

pub struct Logger { /* private fields */ }

A simple logger that splits logging between writing logs that contain only public, non-sensitive content and writing logs that could potentially contain sensitive content.

Writing of potentially sensitive content will be ignored unless the “oak-unsafe” feature is enabled.

Implementations

impl Logger

pub fn new(max_level: LevelFilter) -> Self

Creates a new logger with the specified maximum LevelFilter.

pub fn for_test() -> Self

Creates a new logger for testing using the debug LevelFilter.

pub fn log_sensitive(&self, level: Level, message: &str)

Logs the message at the specified Level, but only if the “oak-unsafe” feature is enabled.

pub fn log_public(&self, level: Level, message: &str)

Logs a message that contains only public, non-sensitive content at the specified Level.

All code that uses this function should be inspected to ensure that the message can never contain any sensitive or non-public information.