Module oak_functions_loader::server

Structs

ExtensionResult

Policy
A policy describing limits on the size of the response and response processing time to avoid side-channel leaks.

ValidatedPolicy
Similar to Policy, but it is guaranteed to be valid. An instance of this can only be created from a valid Policy by calling try_into on it.

WasmHandler

WasmState
WasmState holds runtime values for a particular execution instance of Wasm, handling a single user request. The methods here correspond to the ABI host functions that allow the Wasm module to exchange the request and the response with the Oak functions server. These functions translate values between Wasm linear memory and Rust types.

Constants

ABI_USIZE
Wasm type identifier for position/offset values in linear memory. Any future 64-bit version of Wasm would use a different value.

Traits

OakApiNativeExtension
Trait for implementing extensions, to implement new native functionality.

Functions

apply_policy
Runs the given function and applies the given security policy to the execution of the function and the response returned from it. Serializes and returns the response as a binary protobuf-encoded byte array of a constant size.

format_bytes
Converts a binary sequence to a string if it is a valid UTF-8 string, or formats it as a numeric vector of bytes otherwise.

Type Definitions

AbiPointer

AbiPointerOffset

BoxedExtension