Struct oak_functions_loader::server::WasmState

pub struct WasmState { /* private fields */ }

WasmState holds runtime values for a particular execution instance of Wasm, handling a single user request. The methods here correspond to the ABI host functions that allow the Wasm module to exchange the request and the response with the Oak functions server. These functions translate values between Wasm linear memory and Rust types.

Implementations

impl WasmState

pub fn get_memory(&self) -> &MemoryRef

Helper function to get memory.

pub fn read_buffer_from_wasm_memory(
&self,
buf_ptr: AbiPointer,
buf_len: AbiPointerOffset
) -> Result<Vec<u8>, OakStatus>

Reads the buffer starting at address buf_ptr with length buf_len from the Wasm memory.

pub fn read_extension_args(
&self,
buf_ptr: AbiPointer,
buf_len: AbiPointerOffset
) -> Result<Vec<u8>, Error>

Read arguments for extension from memory of Wasm module.

pub fn write_extension_result(
&mut self,
result: Vec<u8>,
result_ptr_ptr: AbiPointer,
result_len_ptr: AbiPointer
) -> Result<(), OakStatus>

pub fn write_buffer_to_wasm_memory(
&self,
source: &[u8],
dest: AbiPointer
) -> Result<(), OakStatus>

Writes the buffer source at the address dest of the Wasm memory, if source fits in the allocated memory.

pub fn write_u32_to_wasm_memory(
&self,
value: u32,
address: AbiPointer
) -> Result<(), OakStatus>

Writes the u32 value at the address of the Wasm memory.

pub fn alloc_and_write_buffer_to_wasm_memory(
&mut self,
buffer: Vec<u8>,
dest_ptr_ptr: AbiPointer,
dest_len_ptr: AbiPointer
) -> Result<(), OakStatus>

Writes the given buffer by allocating buffer.len() Wasm memory and writing the address of the allocated memory to dest_ptr_ptr and the length to dest_len_ptr.

pub fn read_request(
&mut self,
dest_ptr_ptr: AbiPointer,
dest_len_ptr: AbiPointer
) -> Result<(), OakStatus>

Corresponds to the host ABI function read_request.

pub fn write_response(
&mut self,
buf_ptr: AbiPointer,
buf_len: AbiPointerOffset
) -> Result<(), OakStatus>

Corresponds to the host ABI function write_response.

pub fn write_log_message(
&mut self,
buf_ptr: AbiPointer,
buf_len: AbiPointerOffset
) -> Result<(), OakStatus>

Corresponds to the host ABI function write_log_message.

pub fn invoke_extension_with_handle(
&mut self,
handle: AbiExtensionHandle,
request_ptr: AbiPointer,
request_len: AbiPointerOffset,
response_ptr_ptr: AbiPointer,
response_len_ptr: AbiPointer
) -> Result<(), OakStatus>

pub fn alloc(&mut self, len: u32) -> AbiPointer