Trait oak_functions_loader::server::OakApiNativeExtension

pub trait OakApiNativeExtension {
    fn invoke(&mut self, request: Vec<u8>) -> Result<Vec<u8>, OakStatus>;

    fn get_metadata(&self) -> (String, Signature);

    fn terminate(&mut self) -> Result<()>;

    fn get_handle(&self) -> ExtensionHandle;
}

Trait for implementing extensions, to implement new native functionality.

Required methods

fn invoke(&mut self, request: Vec<u8>) -> Result<Vec<u8>, OakStatus>

Invokes the extension with the given request and returns a result. If no result is expected, the result is empty. An error within the extension is reflected in the OakStatus. TODO(#2701): Stop returning OakStatus for errors internal to extensions.

fn get_metadata(&self) -> (String, Signature)

Metadata about this Extension, including the exported host function name, the function’s signature, and the corresponding ExtensionHandle.

fn terminate(&mut self) -> Result<()>

Performs any cleanup or terminating behavior necessary before destroying the WasmState.

fn get_handle(&self) -> ExtensionHandle

Gets the ExtensionHandle for this extension.