Function oak_functions_loader::server::apply_policy

pub async fn apply_policy<F, S>(
    policy: ServerPolicy, 
    function: F
) -> Result<Response> where
    F: Send + 'static + FnOnce() -> S,
    S: Future<Output = Result<Response>> + Send, 

Runs the given function and applies the given security policy to the execution of the function and the response returned from it. Serializes and returns the response as a binary protobuf-encoded byte array of a constant size.

If the execution of the function takes longer than allowed by the given security policy, an error response with status PolicyTimeViolation is returned. If the size of the body in the response returned by the function is larger than allowed by the security policy, the response is discarded and a response with status PolicySizeViolation is returned instead. In all cases, to keep the total size of the returned byte array constant, the body of the response may be padded by a number of trailing 0s before encoding the response as a binary protobuf message. In this case, the length in the response will contain the effective length of the body. This response is guaranteed to comply with the policy’s size restriction.